test: split e2e customconfigs into baseline and extended
