+++ test/e2e/sequential/baseline/podintegrationautoenablement_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package sequentialbaseline
+package baseline
 
 import (
 	"github.com/onsi/ginkgo/v2"
