+++ test/e2e/sequential/baseline/suite_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package sequentialbaseline
+package baseline
 
 import (
 	"context"
