+++ test/e2e/sequential/extended/suite_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package sequentialextended
+package extended
 
 import (
 	"context"
